Finish encryptDecrypt rework, move demo to project root
