+++ .devcontainer/setup.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-cd KeyGenDistributed
-
 sdk_file='qrypt-security-ubuntu.tgz'
 curl -s https://qrypt.azureedge.net/sdk/cpp/v0.9.2/qrypt-security-0.9.2-ubuntu.tgz --output $sdk_file
 tar -zxvf $sdk_file --strip-components=1 -C QryptSecurity
