--- KeyGenDistributed/.vscode/launch.json
@@ -1,57 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "Linux/Mac: Alice - KeyGenDistributed",
-            "type": "lldb",
-            "request": "launch",
-            "program": "${workspaceFolder}/build/KeyGenDistributed",
-            "cwd": "${workspaceFolder}/build",
-            "args": [
-                "--user=alice",
-                "--token=${env:QRYPT_TOKEN}",
-                "--key-type=aes",
-                "--metadata-filename=metadata.bin"
-            ]
-        },
-        {
-            "name": "Linux/Mac: Bob - KeyGenDistributed",
-            "type": "lldb",
-            "request": "launch",
-            "program": "${workspaceFolder}/build/KeyGenDistributed",
-            "cwd": "${workspaceFolder}/build",
-            "args": [
-                "--user=bob",
-                "--token=${env:QRYPT_TOKEN}",
-                "--metadata-filename=metadata.bin"
-            ]
-        },
-        {
-            "name": "Windows: Alice  - KeyGenDistributed",
-            "type": "cppvsdbg",
-            "request": "launch",
-            "console":"integratedTerminal",
-            "program": "${workspaceFolder}/build/Debug/KeyGenDistributed.exe",
-            "cwd": "${workspaceFolder}/build/Debug",
-            "args": [
-                "--user=alice",
-                "--token=${env:QRYPT_TOKEN}",
-                "--key-type=aes",
-                "--metadata-filename=metadata.bin"
-            ]
-        },
-        {
-            "name": "Windows: Bob  - KeyGenDistributed",
-            "type": "cppvsdbg",
-            "request": "launch",
-            "console":"integratedTerminal",
-            "program": "${workspaceFolder}/build/Debug/KeyGenDistributed.exe",
-            "cwd": "${workspaceFolder}/build/Debug",
-            "args": [
-                "--user=bob",
-                "--token=${env:QRYPT_TOKEN}",
-                "--metadata-filename=metadata.bin"
-            ]
-        }
-    ]
-}
