--- KeyGenDistributed/src/common.h
@@ -1,44 +1,0 @@
-#ifndef COMMON_H
-#define COMMON_H
-
-#include <stdexcept>
-#include <map>
-#include <vector>
-#include <string>
-#include <tuple>
-
-using KeyValuePair = std::tuple<std::string, std::string>;
-
-const uint32_t IVLengthInBytes = 12;
-const uint32_t AESKeyLengthInBytes = 32;
-const uint32_t AESKeyWithIVLengthInBytes = AESKeyLengthInBytes + IVLengthInBytes;
-const uint32_t AETagSizeInBytes = 2;
-constexpr int OPENSSL_SUCCESS = 1;
-constexpr int BMP_HEADER_SIZE = 54;
-
-struct BitmapData {
-    std::vector<uint8_t> header;
-    std::vector<uint8_t> body;
-};
-
-std::vector<uint8_t> readFromFile(const std::string filename);
-
-void writeToFile(const std::string filename, std::vector<uint8_t>& buffer);
-
-void writeToFile(const std::string filename, std::string &buffer);
-
-BitmapData readBitmap(const std::string filename);
-
-void writeBitmap(const std::string filename, const BitmapData& bitmapData);
-
-std::vector<uint8_t> xorVectors(const std::vector<uint8_t> otp, const std::vector<uint8_t> &data);
-
-std::string byteVecToHexStr(std::vector<uint8_t>& bytes);
-
-size_t hexCharToInt(char input);
-
-std::vector<uint8_t> hexStrToByteVec(std::string& str);
-
-std::string toUpper(std::string str);
-
-#endif
