+++ test/MetadataValidityTests.cpp
